+++ app/packages/registration-schema/schema-block-group.ts
@@ -1,6 +1,7 @@
 import { SchemaBlock } from 'ember-osf-web/packages/registration-schema';
 
 export interface SchemaBlockGroup {
+    inputType?: string | null;
     labelBlock?: SchemaBlock;
     inputBlock?: SchemaBlock;
     optionBlocks?: SchemaBlock[];