+++ mirage/factories/institution.ts
@@ -1,11 +1,12 @@
-import { Factory, faker, Trait, trait } from 'ember-cli-mirage';
+import { Factory, faker, trait, Trait } from 'ember-cli-mirage';
 
 import Institution from 'ember-osf-web/models/institution';
 
-import { randomGravatar } from '../utils';
+import { placekitten, randomGravatar } from '../utils';
 
 export interface InstitutionTraits {
     withInstitutionalUsers: Trait;
+    withStatSummary: Trait;
 }
 
 export default Factory.extend<Institution & InstitutionTraits>({
@@ -17,12 +18,27 @@ export default Factory.extend<Institution & InstitutionTraits>({
     },
     assets() {
         return {
+            banner: placekitten(512, 80),
             logo: randomGravatar(100),
         };
     },
+    currentUserIsAdmin: true,
     withInstitutionalUsers: trait<Institution>({
         afterCreate(institution, server) {
-            server.createList('institutional-user', 10, { institution });
+            server.createList('institutional-user', 15, { institution });
+        },
+    }),
+    withStatSummary: trait<Institution>({
+        statSummary() {
+            return {
+                departments: ['Architecture', 'Biology', 'Psychology'].map(department => ({
+                    name: department,
+                    numUsers: faker.random.number({ min: 0, max: 99 }),
+                })),
+                ssoUsersConnected: faker.random.number({ min: 0, max: 999 }),
+                numPrivateProjects: faker.random.number({ min: 0, max: 999 }),
+                numPublicProjects: faker.random.number({ min: 0, max: 999 }),
+            };
         },
     }),
 });